% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealCI.R
\name{arealCI}
\alias{arealCI}
\title{Estimate Confidence Intervals (CI) for Areal Probability Field Via Simulation}
\usage{
arealCI(field, modelFit, polygonList = NULL, rWidth = NULL,
  draws = 1000)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{modelFit}{fitted model object from the field}

\item{polygonList}{list, list of polygons to sample from}

\item{rWidth}{integer, instead of using a polygon list divide the original
bounding box into approximately equal sized squares with rWidth number
squares along the x axis.}

\item{draws}{int, number of draws for simulation to calculate CI's}
}
\value{
data.frame, mean, CI, and sd of areal fields.
}
\description{
Calculate the mean, 95% confidence intervals, and the standard
deviation of areal units from an estimated probability field from a model.
}
\examples{
\dontrun{

unitSim <- simField(
N = 200, rangeE = .7,
offset = c(0.1, 0.2), 
max.edge = c(0.1,0.2),
beta0 = -2,
betaList = list(
    list(type="random", value=2),
    list(type="spatial", value=-.5),
    list(type="cluster", value=-2)
))

pointDF <- samplePoints(unitSim, 500, 100)

unitFit <- runFieldModel(unitSim, pointDF)

head(simulateFieldCI(unitSim, unitFit, rWidth=3))

}

}
