% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbCoords.R
\name{bbCoords}
\alias{bbCoords}
\title{Extract Boundary of Polygon as a Spatial Points Object}
\usage{
bbCoords(SpatPolygon)
}
\arguments{
\item{SpatPolygon}{Spatial Polygons object}
}
\value{
Spatial Points object of length 4
}
\description{
Extracts a bounding box object taken from a spatial polygons
data frame and returns a spatial points object with four coordinates
representing the bounding box.
}
\examples{
require(ar.matrix)
bbCoords(US.df)

}
