% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModelInputs.R
\name{buildModelInputs}
\alias{buildModelInputs}
\title{Build model inputs for use with runFieldModel}
\usage{
buildModelInputs(field, pointDF = NULL, polyDF = NULL, model = TRUE,
  moption = 0)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{pointDF}{data simulated from samplePoints}

\item{polyDF}{data simulated from samplePolygns}

\item{model}{logical, build model inputs for modeling rather than prediction}

\item{moption}{int, intger indicating how polygon data should be estimated 0
is by Reimann sum approximation, 1 is by redistribution, and 2 is by Utazi
approach.}
}
\value{
List of data and parameter objects to be used in TMB model run.
}
\description{
Build model inputs for use with runFieldModel
}
