% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividePolygon.R
\name{dividePolygon}
\alias{dividePolygon}
\title{TDivide field into sections}
\usage{
dividePolygon(shape, rWidth)
}
\arguments{
\item{shape}{spatial polygons data frame}

\item{rWidth}{integer, instead of using a polygon list divide the original
bounding box into approximately equal sized squares with rWidth number 
squares along the x axis.}
}
\value{
spatial polygons data frame section into apprximately square sections
}
\description{
Takes a field object and divides it into rWidth square sections.
}
\examples{
require(sp)
require(ar.matrix)

set.seed(123)
usSim <- simField(
    N = 300,
    shape = US.df,
    rangeE = 3,
    offset = c(1, 2),
    max.edge = c(.25, 1))

plot(dividePolygon(usSim$bound, 4))
plot(dividePolygon(usSim$bound, 10))
plot(dividePolygon(usSim$bound, 16))

}
