% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest.R
\name{find_closest}
\alias{find_closest}
\title{Find index of closest neighbor}
\usage{
find_closest(x1, x2, fullDF)
}
\arguments{
\item{x1}{matrix of long lats for desired match}

\item{x2}{matrix of long lats for matching to}

\item{fullDF}{DF indicating the full id names}
}
\value{
integer vector indicating closest neighbor
}
\description{
cpp code for finding closest neighbor
}
