% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggField.R
\name{ggField}
\alias{ggField}
\title{Run a Model and Estimate Underlying Field From Point and Polygon Data}
\usage{
ggField(field, plotVar = "theta")
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{plotVar}{character, which variable to plot}
}
\value{
ggplot object of underlying field
}
\description{
Run a model and estimate underlying latent and probability
field from point and polygon data using TMB.
}
\examples{
unitSim <- simField(
N = 100, rangeE = .5,
offset = c(0.1, 0.2), 
max.edge = c(0.1,0.2),
beta0 = -2)

ggField(unitSim)
  
}
