% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFieldEst.R
\name{ggFieldEst}
\alias{ggFieldEst}
\title{Run a Model and Estimate Underlying Field From Point and Polygon Data}
\usage{
ggFieldEst(field, predList, sd = FALSE)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{predList}{list of model predictions}

\item{sd}{boolean, plot the standard error of field estimates}
}
\value{
ggplot object of underlying field and model estimates
}
\description{
Run a model and estimate underlying latent and probability
field from point and polygon data using TMB.
}
\examples{
\dontrun{
unitSim <- simField(
N = 100, rangeE = .5,
offset = c(0.1, 0.2), 
max.edge = c(0.1,0.2),
beta0 = -2)

pointDF1 <- samplePoints(unitSim, N=1200, M=100) 
pointDF2 <- samplePoints(unitSim, N=1200, M=100)

modelPredList <- lapply(list(pointDF1, pointDF2), function(x){
    simulateFieldCI(unitSim, runFieldModel(unitSim, x))
})

ggFieldEst(unitSim, modelPredList)
}

}
