% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hRatio.R
\name{hRatio}
\alias{hRatio}
\title{Calculate the height to width ratio of a spatial polygon}
\usage{
hRatio(SpatPolygon)
}
\arguments{
\item{SpatPolygon}{Spatial Polygons object}
}
\value{
numeric, ratio of diff
}
\description{
Calculate the height to width ratio of a spatial polygon
}
\examples{
require(ar.matrix)
hRatio(US.df)

}
