% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDataUtazi.R
\name{runDataUtazi}
\alias{runDataUtazi}
\title{Run a Model and Estimate Underlying Field From Point and Polygon Data for the
Utazi model.}
\usage{
runDataUtazi(field, pointDF, polyDF_, shape3, fullDF,
  timeStructured = FALSE)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{pointDF}{data simulated from samplePoints}

\item{polyDF_}{data simulated from samplePolygns}

\item{shape3}{shape to build adjaceny matrix}

\item{fullDF}{id to xy location and urbanicity status}

\item{timeStructured}{bool add time structured effects}
}
\value{
Model fit object
}
\description{
Run the utazi model on the actual DR data
}
