% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFieldModel.R
\name{runFieldModel}
\alias{runFieldModel}
\title{Run a Model and Estimate Underlying Field From Point and Polygon Data}
\usage{
runFieldModel(field, pointDF = NULL, polyDF = NULL, moption = 0,
  verbose = FALSE, symbolic = TRUE, control = list(eval.max = 10000,
  iter.max = 10000))
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{pointDF}{data simulated from samplePoints}

\item{polyDF}{data simulated from samplePolygns}

\item{moption}{int, intger indicating how polygon data should be estimated 0
is by Reimann sum approximation, 1 is by redistribution, and 2 is by Utazi
approach.}

\item{verbose}{logical, print model fitting information}

\item{symbolic}{logical, use metas reordering in model fitting}

\item{control}{list, control list passed to nlminb}
}
\value{
List of fitted model objects.
}
\description{
Run a model and estimate underlying latent and probability
field from point and polygon data using TMB.
}
\examples{
\dontrun{
unitSim <- simField(
N = 500, rangeE = .7,
offset = c(0.1, 0.2), 
max.edge = c(0.1,0.2),
beta0 = -2,
betaList = list(
    list(type="random", value=2),
    list(type="spatial", value=-.5),
    list(type="cluster", value=-2)
))

pointDF <- samplePoints(unitSim, 500, 100)

runFieldModel(unitSim, pointDF)

}

}
