% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFieldModelUtazi.R
\name{runFieldModelUtazi}
\alias{runFieldModelUtazi}
\title{Run a Model and Estimate Underlying Field From Point and Polygon Data}
\usage{
runFieldModelUtazi(field, pointDF = NULL, polyDF = NULL, moption = 0,
  verbose = FALSE, symbolic = TRUE, control = list(eval.max = 10000,
  iter.max = 10000), rWidth = NULL, aggData = TRUE, shape3 = NULL)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{pointDF}{data simulated from samplePoints}

\item{polyDF}{data simulated from samplePolygns}

\item{moption}{int, intger indicating how polygon data should be estimated 0
approach.}

\item{verbose}{logical, print model fitting information}

\item{symbolic}{logical, use metas reordering in model fitting}

\item{control}{list, control list passed to nlminb}

\item{rWidth}{int, only used in moption 2 to build besag prior}

\item{aggData}{bool, whether to aggregate data as passed in}

\item{shape3}{the shape to build the adjaceny matrix with}
}
\value{
List of fitted model objects.
}
\description{
Run a model and estimate underlying latent and probability
field from point and polygon data using INLA and Utazi method.
}
\examples{
\dontrun{
unitSim <- simField(
N = 500, rangeE = .7,
offset = c(0.1, 0.2), 
max.edge = c(0.1,0.2),
beta0 = -2,
betaList = list(
    list(type="random", value=2),
    list(type="spatial", value=-.5),
    list(type="cluster", value=-2)
))

polyDF <- samplePolygons(unitSim, round(1200*100/9), rWidth=3)

runFieldModelUtazi(unitSim, polyDF=polyDF, moption=2, rWidth=3)

}

}
