% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePoints.R
\name{samplePoints}
\alias{samplePoints}
\title{Take Point Observations from a spatial field.}
\usage{
samplePoints(field, N, M, replace = TRUE, ...)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{N}{int 1, Number of points to simulate}

\item{M}{int, Number of trials for each point either 1 or N long}

\item{replace}{logical, replace pixel probabilities in sampling?}

\item{...}{further argumnets for compatibility}
}
\value{
data.frame with observation number, number of trials, and the id
of the pixel that the trail took place in.
}
\description{
Takes binomial trials from a probability field created by 
simField randomly from points on the field.
}
\examples{

set.seed(123)
unitSim <- simField(
    N = 300, 
    offset = c(0.1, 0.2), 
    max.edge = c(0.05,0.2),
    beta0 = -2,
    betaList = list(
        list(type="random", value=2),
        list(type="spatial", value=-.5),
        list(type="cluster", value=-2)
    ))

samplePoints(unitSim, N=20, M=100) 

}
