% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePPMix.R
\name{samplePPMix}
\alias{samplePPMix}
\title{Take Observations from Points and Polygons within the latent field}
\usage{
samplePPMix(field, N, M, p = 0.5, polygonList = NULL, rWidth = NULL,
  replace = TRUE, ...)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{N}{int, Number of points sampled for each polygon}

\item{M}{int, Number of trials for each point}

\item{p}{numeric >0 & <=1, percentage of polygons sampled}

\item{polygonList}{list, list of polygons to sample from}

\item{rWidth}{integer, instead of using a polygon list divide the original
bounding box into approximately equal sized squares with rWidth number 
squares along the x axis.}

\item{replace}{logical, replace values in the sampling}

\item{...}{further argumnets for compatibility}
}
\value{
list of data.frames with observation number, number of trials, 
and the id of the pixel that the trail took place in.
}
\description{
Takes binomial trials from a probability field created by 
simField randomly from the field where each observation takes place at the 
point level and then p% polygons are removed of their point data and
replaced with representative polygon data.
}
\examples{
require(sp)

set.seed(123)
unitSim <- simField(
    N = 300, 
    offset = c(0.1, 0.2), 
    max.edge = c(0.05,0.2),
    beta0 = -2,
    betaList = list(
        list(type="random", value=2),
        list(type="spatial", value=-.5),
        list(type="cluster", value=-2)
    ))

mixSample <- samplePPMix(unitSim, 30, 150, .5, rWidth=3)
head(mixSample$pointDF)
head(mixSample$polyDF)

}
