% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDataFieldCI.R
\name{simulateDataFieldCI}
\alias{simulateDataFieldCI}
\title{Estimate Confidence Intervals (CI) for Probability Field Via Simulation for
data estimated models}
\usage{
simulateDataFieldCI(modelFit, field, draws = 1000)
}
\arguments{
\item{modelFit}{fitted model object from the field}

\item{field}{field object which simulated underlying data}

\item{draws}{int, number of draws for simulation to calculate CI's}
}
\value{
data.frame, mean, CI, and sd of underlying field.
}
\description{
Calculate the mean, 95% confidence intervals, and the standard
deviation of an estimated probability field from a model.
}
