% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFieldCI.R
\name{simulateFieldCI}
\alias{simulateFieldCI}
\title{Estimate Confidence Intervals (CI) for Probability Field Via Simulation}
\usage{
simulateFieldCI(field, modelFit, draws = 1000)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{modelFit}{fitted model object from the field}

\item{draws}{int, number of draws for simulation to calculate CI's}
}
\value{
data.frame, mean, CI, and sd of underlying field.
}
\description{
Calculate the mean, 95% confidence intervals, and the standard
deviation of an estimated probability field from a model.
}
\examples{
\dontrun{

unitSim <- simField(
N = 200, rangeE = .7,
offset = c(0.1, 0.2), 
max.edge = c(0.1,0.2),
beta0 = -2,
betaList = list(
    list(type="random", value=2),
    list(type="spatial", value=-.5),
    list(type="cluster", value=-2)
))

pointDF <- samplePoints(unitSim, 500, 100)

unitFit <- runFieldModel(unitSim, pointDF)

head(simulateFieldCI(unitSim, unitFit))

}

}
