% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFieldModel.R
\name{runFieldModel}
\alias{runFieldModel}
\title{Run a Model and Estimate Underlying Field From Point and Polygon Data}
\usage{
runFieldModel(field, pointDF = NULL, polyDF = NULL, moption = 0,
  verbose = FALSE, symbolic = TRUE, control = list(eval.max = 10000,
  iter.max = 10000), rWidth = NULL, priors = FALSE, mcmc = FALSE,
  AprojPoly = NULL, ...)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{pointDF}{data simulated from samplePoints}

\item{polyDF}{data simulated from samplePolygns}

\item{moption}{int, intger indicating how polygon data should be estimated 0
is by mixed model approximation, 1 is by redistribution, 2 is by Utazi
approach, 3 is by riemman approximation, 4 is ignoring polygon data,
5 is if all data was known.}

\item{verbose}{logical, print model fitting information}

\item{symbolic}{logical, use metas reordering in model fitting}

\item{control}{list, control list passed to nlminb}

\item{rWidth}{int, only used in moption 2 to build besag prior}

\item{priors}{logical, default FALSE Should priors be evaluated for top level
parameters.}

\item{mcmc}{logical, default FALSE Should model be fit with MCMC. Not
compatible with moption 2.}

\item{AprojPoly}{sparseMatrix, sparse matrix with population weight information for polygons.}

\item{...}{Further arguments to pass to tmbstan}
}
\value{
List of fitted model objects.
}
\description{
Run a model and estimate underlying latent and probability
field from point and polygon data using TMB.
}
\examples{
\dontrun{
unitSim <- simField(
N = 500, rangeE = .7,
offset = c(0.1, 0.2), 
max.edge = c(0.1,0.2),
beta0 = -2,
betaList = list(
    list(type="random", value=2),
    list(type="spatial", value=-.5),
    list(type="cluster", value=-2)
))

pointDF <- samplePoints(unitSim, 500, 100)

runFieldModel(unitSim, pointDF)

}

}
