% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePolygons.R
\name{samplePolygons}
\alias{samplePolygons}
\title{Take Observations from Polygons within the latent field}
\usage{
samplePolygons(field, M, p = 1, polygonList = NULL, rWidth = NULL, ...)
}
\arguments{
\item{field}{field object which simulated underlying data}

\item{M}{int, Number of trails for each polygon}

\item{p}{numeric >0 & <=1, percentage of polygons sampled}

\item{polygonList}{list, list of polygons to sample from}

\item{rWidth}{integer, instead of using a polygon list divide the original
bounding box into approximately equal sized squares with rWidth number 
squares along the x axis.}

\item{...}{further argumnets for compatibility}
}
\value{
data.frame with observation number, number of trails, and the id
of the pixel that the trail took place in.
}
\description{
Takes binomial trails from a probability field created by 
simField randomly from the field where each observation takes place within
a given polygon and is representative of that polygon.
}
\examples{
require(sp)

set.seed(123)
unitSim <- simField(
    N = 300, 
    offset = c(0.1, 0.2), 
    max.edge = c(0.05,0.2),
    beta0 = -2,
    betaList = list(
        list(type="random", value=2),
        list(type="spatial", value=-.5),
        list(type="cluster", value=-2)
    ))

samplePolygons(unitSim, round(1200*100/9), rWidth=3)

}
