% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arealDataCI.R, R/simulateDataFieldCI.R
\name{simulateDataFieldCI}
\alias{simulateDataFieldCI}
\title{Estimate Confidence Intervals (CI) for Probability Field Via Simulation for
data estimated models}
\usage{
simulateDataFieldCI(modelFit, field, draws = 1000, agg5q0 = FALSE)

simulateDataFieldCI(modelFit, field, draws = 1000, agg5q0 = FALSE)
}
\arguments{
\item{modelFit}{fitted model object from the field}

\item{field}{field object which simulated underlying data}

\item{draws}{int, number of draws for simulation to calculate CI's}

\item{agg5q0}{logical, collpase age groups to 5q0?}

\item{popDF}{data.frame, use population weights when aggrgating data}

\item{field}{field object which simulated underlying data}

\item{modelFit}{fitted model object from the field}

\item{draws}{int, number of draws for simulation to calculate CI's}

\item{agg5q0}{logical, collpase age groups to 5q0?}
}
\value{
data.frame, mean, CI, and sd of underlying field.

data.frame, mean, CI, and sd of underlying field.
}
\description{
Calculate the mean, 95% confidence intervals, and the standard
deviation of an estimated probability field from a model.

Calculate the mean, 95% confidence intervals, and the standard
deviation of an estimated probability field from a model.
}
